+++ src/ParsedResponse.php
@@ -13,37 +13,37 @@ interface ParsedResponse
      * Get message body schema
      * @return string
      */
-    public function getSchema();
+    public function getSchema() : string;
 
     /**
      * Set message body schema
      * @param string $schema
      * @return ParsedResponse
      */
-    public function setSchema($schema);
+    public function setSchema(string $schema) : ParsedResponse;
 
     /**
      * Get HTTP status code
      * @return int
      */
-    public function getStatusCode();
+    public function getStatusCode() : int;
 
     /**
      * Set HTTP status code
      * @param int $statusCode
      * @return ParsedResponse
      */
-    public function setStatusCode($statusCode);
+    public function setStatusCode(int $statusCode) : ParsedResponse;
 
     /**
      * Get response body
-     * @return string
+     * @return mixed|string|null
      */
     public function getBody();
 
     /**
      * Set the response body
-     * @param string $body
+     * @param mixed|string|null $body
      * @return ParsedResponse
      */
     public function setBody($body);
@@ -52,12 +52,12 @@ interface ParsedResponse
      * Get header bag
      * @return HeaderBag
      */
-    public function getHeaders();
+    public function getHeaders() : HeaderBag;
 
     /**
      * Replace entire header bag
      * @param HeaderBag $headers
      * @return ParsedResponse
      */
-    public function setHeaders(HeaderBag $headers);
+    public function setHeaders(HeaderBag $headers) : ParsedResponse;
 }